text_client.py: Don't throw exception on missing conf
If we are missing the ".mycroft_cli.conf" file we print a message to the
user informing them that we are ignoring the missing file, but then
throw an exception. Remove the exception throwing and instead allow the
user to continue without the file, as we tell them we are doing.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>