{start,stop}-mycroft.sh: allow system-wide setups
In case of Linux distributions, MyCroft will be installed system-wide
and a git checkout is not available. Without this change,
start-mycroft.sh and stop-mycroft.sh however expect that git checkout to
be available and it won't launch the services without it.

Modify the script to check if it's being ran systemwide, and if so skip
the usual checks and steps that come to play when running in a git
checkout. This way these scripts can be used to start and stop MyCroft
services from distribution installed packages, assuming these scripts
are installed as `start-mycroft` and `stop-mycroft`.