Use unique paths for temp file storage
When transferring the Allure report and the Mycroft logs to the
report host, the zip files were being written to a common
directory. In the event that multiple jobs were writing to or
reading from this directory at the same time, conflicts could
occur.

This ensures that both zip files are written to unique paths
and cleaned up afterward.